Remove gradient and add :target styling
The gradient can show up during loading, which is annoying.

The bottom margin (and the gradient, at the same time) was
added to solve the problem of it not being clear what the
:target was in multipage/ when it was near the bottom of
the page. Instead add a fading background to :target for
`dfn`, `a` and `i` elements.

Also remove invisible text-shadow on headings.