Rename index to index.html on whatwg.org
The many files on whatwg.org with no file ending require a bunch of
fiddling with content-type. Most of the URLs need to be preserved so
that renaming would require redirecting, but not so for these index
files.

If it turns out that any links are broken, redirecting foo/index to foo/
everywhere wouldn't be hard.

Note that https://whatwg.org/specs/web-workers/current-work/ (without
index.html) is redirected but
https://whatwg.org/specs/web-workers/current-work/rationale.html is not.