Retry most curl requests as they sometimes timeout
We are not a 100% sure what exactly causes them to timeout as locally they always seem available around the same time, but hopefully this reduces the fail rate of PRs a bit.

Note that this does not address curl'ing of the deploy.sh script itself and also does not touch curl'ing the HTML checker as that code is likely to change anyway.

Fixes #117 (the remainder).