Do not transform text in .category-list
As far as I know this only affects the HTML Standard.

This makes "MathML math" appear as "mathml math". And "Hidden state" as "hidden state", despite everywhere else using the former. The only somewhat intentional lowercase might be "Text", but we should just lowercase that in the source in that case.