Update syntax highlighting styles
The main purpose of this commit is to adapt to the new markup style in https://github.com/whatwg/html/pull/3768. This also fixes minor inconsistencies and redundancies noted throughout the stylesheet for pre elements. (Most notably, it will treat Bikeshed-produced IDL and CSS blocks the same as those in HTML.) And it removes special treatment of the "asn" class, which does not appear in any current specs.