Output Bikeshed errors if they occur
This is a revision of fdf8cedf0ddc45af217f192d2e967249478b460a which should work correctly. It uses the Accept header to get the right output. It also fixes a bug with that commit where we accidentally passed through the output filename twice. And, it generally tidies up some curl usage in the rest of the script.

Closes #227. Closes #223.