Further changes to margins and warnings
This is a follow-up to 36b8b4571e480ea54938052f7335fc0b9f3f0cee. It
reduces the margins on the various special paragraphs, from 2em to 1.5em
default/1em in lists, with a +0.5em for the annotation-having special
paragraphs (domintro, note, warning, example). This better matches
non-special paragraphs.

Additionally, this adds a very-light-yellow background to the warning
class, as it was an exception in having no background and that caused it
to look strange when it was given the same margins as the
background-having ones.

Finally, this removes some files I was using for local testing, which
accidentally got pushed and committed.