Require complete responses
The individual messages found by `git grep -Li '</html>'` were deleted
and the script run again to get an earlier complete entry where available.

The files deleted in this commit often still have enough information
to identify which message it was, in particular an In-Reply-To value
in mailto: links that are the Message-Id. These might be useful later,
and can be recovered from the parent commit.